+++ GraphSharp/Common/Interfaces/IWeighted.cs
@@ -1,5 +1,5 @@
 namespace GraphSharp.Common;
 public interface IWeighted
 {
-    float Weight { get; set; }
+    double Weight { get; set; }
 }
