+++ GraphSharp/GraphStructures/GraphOperations/FindLocalClusteringCoefficients.cs
@@ -13,14 +13,14 @@ where TEdge : IEdge
     /// Finds local clustering coefficients
     /// </summary>
     /// <returns>Array, where index is node Id and value is coefficient. -1 means this node was not present in the graph.</returns>
-    public float[] FindLocalClusteringCoefficients()
+    public double[] FindLocalClusteringCoefficients()
     {
         //for each node n take it's neighbors
         //induce graph on {n+neighborhoods} in sum N nodes
         //in induced graph find count of edges = K
         //set coeff for n equal K/(N(N-1))
 
-        var coeff = new float[Nodes.MaxNodeId+1];
+        var coeff = new double[Nodes.MaxNodeId+1];
         Array.Fill(coeff,-1f);
         Parallel.ForEach(Nodes,n=>{
             var toInduce = Edges.Neighbors(n.Id).Append(n.Id).ToArray();
@@ -29,8 +29,8 @@ where TEdge : IEdge
                 return;
             }
             var induced = Induce(toInduce);
-            float N = toInduce.Length;
-            float K = induced.Edges.Count;
+            double N = toInduce.Length;
+            double K = induced.Edges.Count;
             coeff[n.Id] = K/(N*(N-1));
         });
         return coeff;
