+++ GraphSharp/GraphStructures/GraphOperations/FindShortestPaths.cs
@@ -13,7 +13,7 @@ where TEdge : IEdge
     /// <param name="nodeId"></param>
     /// <param name="getWeight">When null shortest path is computed by comparing weights of the edges. If you need to change this behavior specify this delegate. Beware that this method will be called in concurrent context and must be thread safe.</param>
     /// <returns>DijkstrasAlgorithm instance that can be used to get path to any other node and length of this path</returns>
-    public DijkstrasAlgorithm<TNode, TEdge> FindShortestPathsDijkstra(int nodeId, Func<TEdge, float>? getWeight = null)
+    public DijkstrasAlgorithm<TNode, TEdge> FindShortestPathsDijkstra(int nodeId, Func<TEdge, double>? getWeight = null)
     {
         return FindShortestPathsDijkstraBase(
             nodeId,
@@ -26,14 +26,14 @@ where TEdge : IEdge
     /// <param name="nodeId"></param>
     /// <param name="getWeight">When null shortest path is computed by comparing weights of the edges. If you need to change this behavior specify this delegate. Beware that this method will be called in concurrent context and must be thread safe.</param>
     /// <returns>DijkstrasAlgorithm instance that can be used to get path to any other node and length of this path</returns>
-    public DijkstrasAlgorithm<TNode, TEdge> FindShortestPathsParallelDijkstra(int nodeId, Func<TEdge, float>? getWeight = null)
+    public DijkstrasAlgorithm<TNode, TEdge> FindShortestPathsParallelDijkstra(int nodeId, Func<TEdge, double>? getWeight = null)
     {
         return FindShortestPathsDijkstraBase(
             nodeId,
             pathFinder=>GetParallelPropagator(pathFinder),
             getWeight);
     }
-    DijkstrasAlgorithm<TNode, TEdge> FindShortestPathsDijkstraBase(int nodeId,Func<DijkstrasAlgorithm<TNode, TEdge>,IPropagator<TNode,TEdge>> createPropagator, Func<TEdge, float>? getWeight = null)
+    DijkstrasAlgorithm<TNode, TEdge> FindShortestPathsDijkstraBase(int nodeId,Func<DijkstrasAlgorithm<TNode, TEdge>,IPropagator<TNode,TEdge>> createPropagator, Func<TEdge, double>? getWeight = null)
     {
         var pathFinder = new DijkstrasAlgorithm<TNode, TEdge>(nodeId, StructureBase, getWeight);
         var propagator = createPropagator(pathFinder);
